+++ docs/locales/en/LC_MESSAGES/development/build_kuscia_cn.po
@@ -244,16 +244,16 @@ msgstr "Add New Test Cases"
 
 #: ../../development/build_kuscia_cn.md:132
 msgid ""
-"如果您希望为 Kuscia 新增更多的测试用例，您可以在 Kuscia 项目的 `scripts/test/suite/center` 和 "
-"`scripts/test/suite/center/p2p` 下添加您的测试用例代码。 您可以参考 "
+"如果您希望为 Kuscia 新增更多的测试用例，您可以在 Kuscia 项目的 `tests/suite/center` 和 "
+"`tests/suite/center/p2p` 下添加您的测试用例代码。 您可以参考 "
 "`scripts/test/suite/center/basic.sh` 和 "
 "`scripts/test/suite/center/example.sh` 来编写您的测试用例。 Kuscia 使用 "
 "[shunit2](https://github.com/kward/shunit2) 作为测试框架，安装在 "
 "`scripts/test/vendor` 下，您可以使用其中的断言函数。 Kuscia 也准备了一些常用的函数，您可以在 "
 "`scripts/test/suite/core` 下找到。"
 msgstr ""
-"If you want to add more test cases for Kuscia, you can add your test case code under `scripts/test/suite/center` and "
-"`scripts/test/suite/center/p2p` in the Kuscia project. You can refer to "
+"If you want to add more test cases for Kuscia, you can add your test case code under `tests/suite/center` and "
+"`tests/suite/center/p2p` in the Kuscia project. You can refer to "
 "`scripts/test/suite/center/basic.sh` and "
 "`scripts/test/suite/center/example.sh` to write your test cases. Kuscia uses "
 "[shunit2](https://github.com/kward/shunit2) as the testing framework, installed under "
@@ -286,11 +286,11 @@ msgstr "Add executable permissions to your test suite file: `chmod a+x {YOUR_TES
 
 #: ../../development/build_kuscia_cn.md:142
 msgid ""
-"在 `scripts/test/integration_test.sh` 文件中注册您的测试用例集。如 "
+"在 `tests/integration_test.sh` 文件中注册您的测试用例集。如 "
 "`TEST_SUITES[\"center.example\"]=\"./test/suite/center/example.sh\"`。变量 "
 "`TEST_SUITES` 的 key 为您的测试用例集的名称。"
 msgstr ""
-"Register your test suite in the `scripts/test/integration_test.sh` file. For example: "
+"Register your test suite in the `tests/integration_test.sh` file. For example: "
 "`TEST_SUITES[\"center.example\"]=\"./test/suite/center/example.sh\"`. The key of the "
 "`TEST_SUITES` variable is the name of your test suite."
