+++ pkg/kusciaapi/service/domaindata_source.go
@@ -634,8 +634,8 @@ func parseAndNormalizeDataSource(sourceType string, info *kusciaapi.DataSourceIn
 		}
 		// truncate slash
 		// fix Issue:
-		// datasource-path: "/home/admin" ; datasource-prefix: "/test/" ; domainData-relativeURI: "/data/alice.csv"
-		// os.path.join(path,prefix,uri) would be /data/alice.csv , this is not expect, expect is /home/admin/test/data/alice.csv
+		// datasource-path: "/home/admin" ; datasource-prefix: "/test/" ; domainData-relativeURI: "/data/alice.csv" -> "/test/data/alice.csv"
+		// os.path.join(path,prefix,uri) would be /data/alice.csv , this is not expect, expect is /home/admin/test/data/alice.csv -> /home/admin/test/data/alice.csv
 		// so trim the prefix filepath.Separator
 		info.Oss.Bucket = strings.Trim(info.Oss.Bucket, string(filepath.Separator))
 		info.Oss.Prefix = strings.Trim(info.Oss.Prefix, string(filepath.Separator))
