+++ tests/suite/README.md
@@ -1,0 +1,1 @@
+# Moved to tests/suite/README.md
