+++ tests/suite/center/base.sh
