+++ tests/suite/center/example.sh
