+++ tests/suite/center/kuscia_images.sh
