+++ tests/suite/core/functions.sh
